+++ minix/include/minix/bdev.h
@@ -16,7 +16,7 @@ extern ssize_t bdev_gather(dev_t dev, u64_t pos, iovec_t *vec, int count,
 	int flags);
 extern ssize_t bdev_scatter(dev_t dev, u64_t pos, iovec_t *vec, int count,
 	int flags);
-extern int bdev_ioctl(dev_t dev, int request, void *buf,
+extern int bdev_ioctl(dev_t dev, unsigned long request, void *buf,
 	endpoint_t user_endpt);
 
 /* Asynchronous API. */
@@ -36,7 +36,7 @@ extern bdev_id_t bdev_gather_asyn(dev_t dev, u64_t pos, iovec_t *vec,
 	int count, int flags, bdev_callback_t callback, bdev_param_t param);
 extern bdev_id_t bdev_scatter_asyn(dev_t dev, u64_t pos, iovec_t *vec,
 	int count, int flags, bdev_callback_t callback, bdev_param_t param);
-extern bdev_id_t bdev_ioctl_asyn(dev_t dev, int request, void *buf,
+extern bdev_id_t bdev_ioctl_asyn(dev_t dev, unsigned long request, void *buf,
 	endpoint_t user_endpt, bdev_callback_t callback, bdev_param_t param);
 
 extern int bdev_wait_asyn(bdev_id_t id);
